requirements.txt
pandas
matplotlib
scikit-learn
python-dotenv
pyTelegramBotAPI
joblib
